requirements.txt
streamlit
cohere
faiss-cpu
pypdf
sentence-transformers
numpy